Prepare to fork or remove Om
I discovered a race condition bug where the Om cursor disagreed with the DOM
about the document. This is a problem because React.js's onChange relies on the
DOM to update the cursor.  This is caused by the batched rendering optimizations
in Om. I will either have to fork Om to disable batched updates or use something
else.